# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Template Flask MongoDB API
  description: |
    A Flask + MongoDB API template that follows patterns established in api_utils.
    This API implements three domains: Control, Create, and Consume.
    
    ## Authentication
    All endpoints (except `/dev-login` and `/metrics`) require JWT authentication via the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```
    
    For local development, obtain a token from the `/dev-login` endpoint.
  version: 0.1.0
  contact:
    name: AgileCraftsPeople
    email: AgileCraftsPeople@gmail.com
  license:
    name: Apache-2.0

servers:
  - url: http://localhost:8184
    description: Local development server

tags:
  - name: Authentication
    description: Development authentication endpoints
  - name: Config
    description: Configuration endpoints
  - name: Metrics
    description: Monitoring endpoints
  - name: Control
    description: Control domain operations (create, read, update)
  - name: Create
    description: Create domain operations (create, read)
  - name: Consume
    description: Consume domain operations (read-only)

paths:
  /dev-login:
    post:
      tags:
        - Authentication
      summary: Get development JWT token
      description: Issues a signed JWT for local/development environments
      operationId: devLogin
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                subject:
                  type: string
                  description: User subject/ID
                  default: dev-user-1
                  example: dev-user-1
                roles:
                  type: array
                  items:
                    type: string
                  description: User roles
                  default: ["developer"]
                  example: ["admin", "developer"]
      responses:
        '200':
          description: Successfully issued JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT access token
                  token_type:
                    type: string
                    example: bearer
                  expires_at:
                    type: string
                    format: date-time
                    description: Token expiration timestamp
                  subject:
                    type: string
                    description: User subject
                  roles:
                    type: array
                    items:
                      type: string
                    description: User roles

  /api/config:
    get:
      tags:
        - Config
      summary: Get configuration
      description: Returns the current configuration as JSON (requires authentication)
      operationId: getConfig
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Configuration retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  config_items:
                    type: array
                    items:
                      type: object
                  versions:
                    type: array
                    items:
                      type: object
                  enumerators:
                    type: array
                    items:
                      type: object
                  token:
                    type: object
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /metrics:
    get:
      tags:
        - Metrics
      summary: Prometheus metrics
      description: Returns Prometheus-formatted metrics
      operationId: getMetrics
      responses:
        '200':
          description: Metrics retrieved successfully
          content:
            text/plain:
              schema:
                type: string

  /api/control:
    get:
      tags:
        - Control
      summary: Get all control documents
      description: Retrieve all control documents, optionally filtered by name
      operationId: getControls
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          description: Optional name filter (supports partial matches, case-insensitive)
          schema:
            type: string
            example: my-control
      responses:
        '200':
          description: Successfully retrieved controls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Control'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Control
      summary: Create a new control document
      description: Create a new control document (requires staff or admin role)
      operationId: createControl
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ControlInput'
      responses:
        '201':
          description: Control created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    pattern: '^[0-9a-fA-F]{24}$'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/control/{controlId}:
    get:
      tags:
        - Control
      summary: Get control by ID
      description: Retrieve a specific control document by its ID
      operationId: getControl
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ObjectId'
      responses:
        '200':
          description: Successfully retrieved control
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Control'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags:
        - Control
      summary: Update control document
      description: Update a control document (requires admin role). Cannot update _id, created, or saved fields.
      operationId: updateControl
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ObjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ControlUpdate'
      responses:
        '200':
          description: Control updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Control'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/create:
    get:
      tags:
        - Create
      summary: Get all create documents
      description: Retrieve all create documents
      operationId: getCreates
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully retrieved creates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Create'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Create
      summary: Create a new create document
      description: Create a new create document (requires staff or admin role)
      operationId: createCreate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInput'
      responses:
        '201':
          description: Create document created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    pattern: '^[0-9a-fA-F]{24}$'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/create/{createId}:
    get:
      tags:
        - Create
      summary: Get create by ID
      description: Retrieve a specific create document by its ID
      operationId: getCreate
      security:
        - bearerAuth: []
      parameters:
        - name: createId
          in: path
          required: true
          description: The create ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Successfully retrieved create
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Create'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/consume:
    get:
      tags:
        - Consume
      summary: Get all consume documents
      description: Retrieve all consume documents (read-only), optionally filtered by name
      operationId: getConsumes
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          description: Optional name filter (supports partial matches, case-insensitive)
          schema:
            type: string
            example: my-consume
      responses:
        '200':
          description: Successfully retrieved consumes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Consume'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/consume/{consumeId}:
    get:
      tags:
        - Consume
      summary: Get consume by ID
      description: Retrieve a specific consume document by its ID (read-only)
      operationId: getConsume
      security:
        - bearerAuth: []
      parameters:
        - name: consumeId
          in: path
          required: true
          description: The consume ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Successfully retrieved consume
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Consume'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from `/dev-login` endpoint.
        Include the token in the Authorization header: `Bearer <token>`

  parameters:
    ObjectId:
      name: controlId
      in: path
      required: true
      description: MongoDB ObjectId
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
        example: 507f1f77bcf86cd799439011

  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: Unauthorized

    Breadcrumb:
      type: object
      description: Audit trail information
      additionalProperties: false
      required:
        - from_ip
        - by_user
        - at_time
        - correlation_id
      properties:
        from_ip:
          type: string
          format: ip
          description: HTTP request remote IP address
          example: 192.168.1.1
        at_time:
          type: string
          format: date-time
          description: Timestamp of the action
          example: "2024-01-15T10:00:00Z"
        by_user:
          type: string
          description: User ID who performed the action
          pattern: '^[^\s]{1,40}$'
          example: user-123
        correlation_id:
          type: string
          description: Logging correlation ID
          pattern: '^[^\s]{1,40}$'
          example: corr-abc123

    Control:
      type: object
      description: Control domain document
      additionalProperties: false
      required:
        - _id
        - name
        - created
        - saved
      properties:
        _id:
          type: string
          description: MongoDB document ID
          pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: Control name
          pattern: '^[^\s]{1,40}$'
          example: my-control
        description:
          type: string
          description: Control description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample control document
        status:
          type: string
          description: Control status
          enum:
            - active
            - archived
          example: active
        created:
          $ref: '#/components/schemas/Breadcrumb'
        saved:
          $ref: '#/components/schemas/Breadcrumb'

    ControlInput:
      type: object
      description: Input for creating a control document
      properties:
        name:
          type: string
          description: Control name
          pattern: '^[^\s]{1,40}$'
          example: my-control
        description:
          type: string
          description: Control description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample control document
        status:
          type: string
          description: Control status
          enum:
            - active
            - archived
          example: active

    ControlUpdate:
      type: object
      description: Fields to update in a control document (cannot include _id, created, saved)
      properties:
        name:
          type: string
          description: Control name
          pattern: '^[^\s]{1,40}$'
        description:
          type: string
          description: Control description
          pattern: '^[^\t\n]{0,255}$'
        status:
          type: string
          description: Control status
          enum:
            - active
            - archived

    Create:
      type: object
      description: Create domain document
      additionalProperties: false
      required:
        - _id
        - name
        - created
      properties:
        _id:
          type: string
          description: MongoDB document ID
          pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: Create name
          pattern: '^[^\s]{1,40}$'
          example: my-create
        description:
          type: string
          description: Create description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample create document
        status:
          type: string
          description: Create status
          example: active
        created:
          $ref: '#/components/schemas/Breadcrumb'

    CreateInput:
      type: object
      description: Input for creating a create document
      properties:
        name:
          type: string
          description: Create name
          pattern: '^[^\s]{1,40}$'
          example: my-create
        description:
          type: string
          description: Create description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample create document
        status:
          type: string
          description: Create status
          example: active

    Consume:
      type: object
      description: Consume domain document (read-only)
      additionalProperties: false
      required:
        - _id
        - name
      properties:
        _id:
          type: string
          description: MongoDB document ID
          pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: Consume name
          pattern: '^[^\s]{1,40}$'
          example: my-consume
        description:
          type: string
          description: Consume description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample consume document
        status:
          type: string
          description: Consume status
          example: active